Themes (#3)
* testing out adding themes

* started to add catppuccin

* It's coming together.

* Changed arrays to pointers.

Changed color_1,2..etc from char arrays to char *. I feel assigning a
pointer the address of the color would be more efficient, instead of all of the strncpy's I was
using before.

Added catppuccin and tokyonight.

* Added kanagawa theme

* Theme functionality pretty much done.

Made the argc statements more robust. Now will behave as expected
whether any args are passed or not and in any order. If no args are
passed it will default to tokyonight theme with no blank line spacer.