Was able to speed it up some.
The zig lang repo and llvm repos were the only ones I have on my system
that are large enough to cause this slow down. After combining three
functions into a signle 'git status --porcelain' call. The zig repo was
very snappy, just as fast as tokyobash repo. But llvm remained the same.
So Im including an option to disable to status bar if the user so wants.