Icons off center in ghostty(source build)
ghostty recently added a bunch of new nerd font rendering code. And now
it seems that alteast some icons are off center and bleed into the next
cell. After testing with my systems default terminal (Konsole) I see
that it was an issue with ghostty. So removed the space that was fixing
an issue that was probrably only affecting me.