iss: false positive.
Fix: cap Day modifier at 15

Found test case with a false positive: If fetchtimer = 30d and month
difference is greater than one. With Febuary as the month between fetch
and current date.

For example, if fetch date is Jan 31. And current date is March 1. It's
techincally over a month since the repo was updated. But in reality,
with Febuary being only 28 days long, its only been 29 days. I'm trying
to think of a simple way to solve this, but for now I'm just going to
limit the max for the Day modifier to 15.