Backgrounds Update.
Each of the 3 themes now have a background that can be enabled via the
config setting `background`.

There is now a `fetch` setting in the config file. You can now have the
statusbar enabled without fetch causing occasional slowdowns on large
repos. Or even when you would go beyond the root dir of the repo would
also cause noticable slowdowns.
So now statusbar is enabled and fetch is disabled by default.

Moved most of the printfs from tokyobash.c to a new lib, printlib.c.

Still need to update the readme with screenshots and updated info.