I broke it.
Last commit compiled and worked fine on my laptop but gave issues on my
desktop. I was afraid I might have issues with colors.h, but hoped with
this being a small project, that it would be ok..

So, I made 2 new structs to hold the color definitions and another for
pointers to the definitions. Now I can access the color definitions from
the printPath functions w/o needing colors.h.

Added 4 functions. One to fill the colors struct with the definition
strings. And another to assign the pointers.
Last two are to print the path with or without a background.

This is still in a broken state, but its functional, hopefully.
Still need to test on my dekstop.