Fix for the 1-3 second slow down when not in the root of a repo.
Finally realized what was causing the slowdown.
I didn't know that the '.git/FETCH_HEAD' part of 'stat .git/FETCH_HEAD'
was a path and not just an argument. So it was hanging since the .git
directory isin't where it thinks it should be.
But it's been fixed and should now correctly find the .git directory.