Efficiency and Colors update.
Complete rework on how the color definitions and pointers worked.
Before, every color was initialized and pointers set depending on the user
settings. Now, only the colors for the chosen theme are initialized.
and pointers where removed since the colorDefs are now used instead.

Reworked a lot of the colors for all 3 themes.
Changed the no-background variants to change their user@host colors when in
root or /mnt like the background variants do. So path stays the same color while
user@host changes.

Removed 'git' config setting.
I originally added it to disable fetching. But with the config settings 'fetch'
and 'statusbar', its no longer needed.