Rewoked a lot of the colors for Catppuccin and Kanagawa.
Also changed the no-background variants to change their user@host colors when in
root or /mnt like the background variants do. So path stays the same color while
user@host changes.

I think this will be the last major color change.
Feel free to leave suggestions or feedback in Discussions or in an email.