Added slash '/' back to end of path if length is 1 and not in root.
Accidentally removed this in a half sleep daze when I thought I was cleaning up
printlib and removing un-needed lines. Thought about keeping it, but I like the
way ~/ looks over just ~ . And incase you're wondering, the check if we are in
root is there so we don't have //, which just looks silly.