Fixed a bug when the path was abbreviated and current dir name is > 22 chars.
The remCurntDir() func was only checking for a '/'. So if your current
dir has a long name, and it got abbreviated as part of abvPath(),
remCurntDir() would have removed upto the first '/' on the left the
abbreviation. But now it will check for either a '/' or three '.'.
Whatever comes first.